feat!: refactor PdfGenerator.java and add some Todo comment
