+++ app/src/main/java/com/example/b07demosummer2024/DBOperation.java
@@ -34,6 +34,8 @@ import java.util.Collections;
 import java.util.regex.Pattern;
 import java.util.regex.Matcher;
 
+import data.StringFilter;
+
 public class DBOperation {
     private DatabaseReference ref;
     private StorageReference mediaRef;
@@ -104,7 +106,6 @@ public class DBOperation {
 
             @Override
             public void onCancelled(@NonNull DatabaseError error) {
-                System.out.println("WTF");
             }
         });
 
@@ -112,6 +113,7 @@ public class DBOperation {
     }
 
     public UploadTask addImage(Uri selectedMedia, AddItemFragment fragment, int lotNumber) {
+        // Todo: check if mediaRef null
         String id = "id" + lotNumber;
         StorageReference media = mediaRef.child(id);
         UploadTask uploadTask = media.putFile(selectedMedia);
@@ -135,28 +137,25 @@ public class DBOperation {
         });
     }
 
-    private String buildRegex(String find) {
-        String [] words = find.toLowerCase().split(" ");
-        StringBuilder regex = new StringBuilder("\\b(");
-        for (String word : words) {
-            regex.append("\\w*").append(word).append("\\w*");
-            if (word != words[words.length - 1]) {
-                regex.append("|");
-            }
+    private boolean isMatchedItem(Item item, Item criteria) {
+        if (criteria.getLotNumber() != null && item.getLotNumber() != null &&
+                !Objects.equals(item.getLotNumber(), criteria.getLotNumber())) {
+            return false;
         }
-        regex.append(")\\b");
-        return regex.toString();
-    }
 
-    private Boolean matchByRegex(String target, String find) {
-        String regex = buildRegex(find);
-        Pattern pattern = Pattern.compile(regex);
-        Matcher matcher = pattern.matcher(target.toLowerCase());
-        while (matcher.find()) {
-            // if there is at least one instance, return True
-            return true;
+        if (!criteria.getName().isEmpty() && !item.getName().isEmpty() &&
+                !StringFilter.matchByRegex(item.getName(), criteria.getName())) {
+            return false;
         }
-        return false;
+        if (!criteria.getCategory().isEmpty() &&
+                !Objects.equals(criteria.getCategory(), item.getCategory())) {
+            return false;
+        }
+        if (!criteria.getPeriod().isEmpty() &&
+                !Objects.equals(criteria.getPeriod(), item.getPeriod())) {
+            return false;
+        }
+        return true;
     }
 
     public Task<List<Item>> searchItem(Item criteria) {
@@ -173,24 +172,9 @@ public class DBOperation {
                         continue;
                     }
 
-                    if (criteria.getLotNumber() != null && item.getLotNumber() != null &&
-                            !Objects.equals(item.getLotNumber(), criteria.getLotNumber())) {
-                        continue;
+                    if (isMatchedItem(item, criteria)) {
+                        filteredItem.add(item);
                     }
-
-                    if (!criteria.getName().isEmpty() && !item.getName().isEmpty() &&
-                            !matchByRegex(item.getName(), criteria.getName())) {
-                        continue;
-                    }
-                    if (!criteria.getCategory().isEmpty() &&
-                            !Objects.equals(criteria.getCategory(), item.getCategory())) {
-                        continue;
-                    }
-                    if (!criteria.getPeriod().isEmpty() &&
-                            !Objects.equals(criteria.getPeriod(), item.getPeriod())) {
-                        continue;
-                    }
-                    filteredItem.add(item);
                 }
 
                 tcs.setResult(filteredItem);
@@ -198,7 +182,6 @@ public class DBOperation {
 
             @Override
             public void onCancelled(@NonNull DatabaseError error) {
-                System.out.println("WTF");
             }
         });
