+++ app/src/main/java/data/PDFGenerator.java
@@ -1,8 +1,7 @@
-package com.example.b07demosummer2024;
+package data;
 import static android.Manifest.permission.READ_EXTERNAL_STORAGE;
 import static android.Manifest.permission.WRITE_EXTERNAL_STORAGE;
 
-import android.app.Activity;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -11,25 +10,19 @@ import android.graphics.Paint;
 import android.graphics.Typeface;
 import android.graphics.pdf.PdfDocument;
 import android.os.Environment;
-import android.util.Log;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
 
-import com.google.android.gms.tasks.OnFailureListener;
+import com.example.b07demosummer2024.Item;
+import com.example.b07demosummer2024.R;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-import com.google.android.gms.tasks.TaskCompletionSource;
-import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 
 import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -38,13 +31,10 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.concurrent.Callable;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.function.BiConsumer;
 
+// Todo: implement interface for this class to conform OCP
 public class PDFGenerator {
     Fragment curFrag;
     private int pageHeight = 500;
@@ -86,12 +76,13 @@ public class PDFGenerator {
                                                  List<Item> list, boolean descImgOnly) {
         CompletableFuture<Void> future = CompletableFuture.completedFuture(null);
         for (Item item : list) {
-            future = future.thenCompose(avoid -> drawItem(pdfDocument, item, descImgOnly));
+            future = future.thenCompose(avoid -> setUpPage(pdfDocument, item, descImgOnly));
         }
         return future;
     }
 
-    private CompletableFuture<Void> drawItem(PdfDocument pdfDocument, Item item,
+    // Todo: handling video file type accordingly
+    private CompletableFuture<Void> setUpPage(PdfDocument pdfDocument, Item item,
                                              boolean descImgOnly) {
         CompletableFuture<Void> future = new CompletableFuture<>();
         ref.child(item.getMediaLink()).getBytes(Long.MAX_VALUE).addOnSuccessListener(new OnSuccessListener<byte[]>() {
@@ -102,28 +93,31 @@ public class PDFGenerator {
                 PdfDocument.Page myPage = pdfDocument.startPage(mypageInfo);
                 Canvas canvas = myPage.getCanvas();
                 InputStream inputStream = new ByteArrayInputStream(bytes);
+
                 bmp = BitmapFactory.decodeStream(inputStream);
                 scaledbmp = Bitmap.createScaledBitmap(bmp, imgWidth, imgHeight, false);
 
-                canvas.drawBitmap(scaledbmp, 30, 80, new Paint());
-                System.out.println(descImgOnly);
-
-                drawText(canvas, item.getName(), pageWidth / 2, 30, 25, true,
-                        true);
-                if (!descImgOnly) {
-                    drawText(canvas, "Period: " + item.getPeriod(), pageWidth / 2, 45,
-                            13, true, true);
-                    drawText(canvas, "Category: " + item.getCategory(), pageWidth / 2, 60,
-                            13, true, true);
-                }
-                drawText(canvas, item.getDescription(), 250, 90, 15,
-                        false, false);
+                drawPage(canvas, item, descImgOnly);
 
                 pdfDocument.finishPage(myPage);
                 future.complete(null);
             }
         });
         return future;
+    }
+
+    private void drawPage(Canvas canvas, Item item, boolean descImgOnly) {
+        canvas.drawBitmap(scaledbmp, 30, 80, new Paint());
+        drawText(canvas, item.getName(), pageWidth / 2, 30, 25, true,
+                true);
+        if (!descImgOnly) {
+            drawText(canvas, "Period: " + item.getPeriod(), pageWidth / 2, 45,
+                    13, true, true);
+            drawText(canvas, "Category: " + item.getCategory(), pageWidth / 2, 60,
+                    13, true, true);
+        }
+        drawText(canvas, item.getDescription(), 250, 90, 15,
+                false, false);
     }
 
     private void drawText(Canvas canvas, String text, int x, int y, int fontSize, boolean isBold,
@@ -147,16 +141,16 @@ public class PDFGenerator {
     private List<String> splitTextIntoLines(String text, int max_width) {
         List<String> strlist = new ArrayList<>();
         String[] words = text.split(" ");
-        String curLine = "";
+        StringBuilder curLine = new StringBuilder();
         for (String word : words) {
             if (curLine.length() + word.length() > max_width) {
-                strlist.add(curLine);
-                curLine = "";
+                strlist.add(curLine.toString());
+                curLine = new StringBuilder();
             }
-            curLine += word + " ";
+            curLine.append(word).append(" ");
         }
-        if (!curLine.isEmpty()) {
-            strlist.add(curLine);
+        if (curLine.length() > 0) {
+            strlist.add(curLine.toString());
         }
         return strlist;
     }
